run a script for proxy election
keep it simple, instead of fiddling around with jinja2 and Ansible yaml syntax
Major drawback is the misbehaviour of | min on a list of strings.
There is no possibility / workaround for converting a list of strings to a list of numbers.
The only viable implementation would involve set_facts iteration which is broken in version 2.5.1